// src/generated/models/Aur_quotesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Aur_quotesaur_quote_status = {
  751820000: 'QuoteWon',
  751820001: 'QuoteLost'
} as const;
export type Aur_quotesaur_quote_status = keyof typeof Aur_quotesaur_quote_status;
export const Aur_quotesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Aur_quotesstatecode = keyof typeof Aur_quotesstatecode;
export const Aur_quotesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Aur_quotesstatuscode = keyof typeof Aur_quotesstatuscode;

export interface Aur_quotesBase {
  aur_ai_generated_summary?: string;
  "aur_benefits@odata.bind"?: string;
  "aur_coverage@odata.bind"?: string;
  "aur_deductibles@odata.bind"?: string;
  "aur_exclusions@odata.bind"?: string;
  aur_gross_premium?: number;
  "aur_inclusions@odata.bind"?: string;
  aur_loading_premium?: number;
  aur_name: string;
  "aur_plan@odata.bind"?: string;
  "aur_product@odata.bind"?: string;
  aur_quote_status?: Aur_quotesaur_quote_status;
  aur_quoteid: string;
  "aur_quotes@odata.bind"?: string;
  aur_reason?: string;
  aur_total_premium?: number;
  aur_vat?: number;
  "aur_warranties@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Aur_quotesstatecode;
  statuscode?: Aur_quotesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Aur_quotes extends Aur_quotesBase {
  aur_benefitsname?: string;
  aur_coveragename?: string;
  aur_deductiblesname?: string;
  aur_exclusionsname?: string;
  aur_inclusionsname?: string;
  aur_planname?: string;
  aur_productname?: string;
  aur_quote_statusname?: string;
  aur_quotesname?: string;
  aur_warrantiesname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  aur_benefits?: object;
  _aur_benefits_value?: string;
  aur_coverage?: object;
  _aur_coverage_value?: string;
  aur_deductibles?: object;
  _aur_deductibles_value?: string;
  aur_exclusions?: object;
  _aur_exclusions_value?: string;
  aur_inclusions?: object;
  _aur_inclusions_value?: string;
  aur_plan?: object;
  _aur_plan_value?: string;
  aur_product?: object;
  _aur_product_value?: string;
  aur_quotes?: object;
  _aur_quotes_value?: string;
  aur_warranties?: object;
  _aur_warranties_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}