// src/generated/models/Aur_productsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Aur_productsesaur_product_status = {
  751820000: 'Draft',
  751820001: 'Publish'
} as const;
export type Aur_productsesaur_product_status = keyof typeof Aur_productsesaur_product_status;
export const Aur_productsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Aur_productsesstatecode = keyof typeof Aur_productsesstatecode;
export const Aur_productsesstatuscode = {
  1: 'Draft',
  2: 'Inactive',
  751820001: 'Publish'
} as const;
export type Aur_productsesstatuscode = keyof typeof Aur_productsesstatuscode;

export interface Aur_productsesBase {
  aur_buy_button?: string;
  aur_buy_heading?: string;
  aur_details?: string;
  aur_heading?: string;
  aur_incomingemailaddress?: string;
  aur_name: string;
  aur_order?: string;
  aur_premium?: string;
  aur_product_name_ar?: string;
  aur_product_status?: Aur_productsesaur_product_status;
  aur_productsid: string;
  aur_remarks?: string;
  aur_short_details?: string;
  aur_slogan?: string;
  aur_terms_conditions?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Aur_productsesstatecode;
  statuscode?: Aur_productsesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Aur_productses extends Aur_productsesBase {
  aur_product_statusname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}