// src/generated/models/Aur_quotes_detailsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Aur_quotes_detailsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Aur_quotes_detailsesstatecode = keyof typeof Aur_quotes_detailsesstatecode;
export const Aur_quotes_detailsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Aur_quotes_detailsesstatuscode = keyof typeof Aur_quotes_detailsesstatuscode;

export interface Aur_quotes_detailsesBase {
  "aur_business_rules@odata.bind"?: string;
  aur_condition_met?: string;
  aur_evidence?: string;
  aur_name?: string;
  "aur_quotes@odata.bind"?: string;
  aur_quotes_detailsid: string;
  aur_response?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Aur_quotes_detailsesstatecode;
  statuscode?: Aur_quotes_detailsesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Aur_quotes_detailses extends Aur_quotes_detailsesBase {
  aur_business_rulesname?: string;
  aur_quotesname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  aur_business_rules?: object;
  _aur_business_rules_value?: string;
  aur_quotes?: object;
  _aur_quotes_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}